need to work

+++ WeatherAppTest/DataModels.swift
@@ -16,8 +16,8 @@ struct Weather: Codable {
 
 struct Main: Codable  {
     var temp: Double = 0.0
-    var humidity: Int = 0
-    var pressure: Int = 0
+    var humidity: Double = 0
+    var pressure: Double = 0
 }
 
 struct WeatherData: Codable  {
